Added Type Checking and Casting Logic

+++ A4/tests/test01.c
@@ -1,7 +1,7 @@
 /* test01.c - Basic arithmetic and assignment */
 int a = 5;
 int b = 10;
-int c = a + b;
+float c = a + b;
 int d = a * b - c;
 int e;
 e = d / 2;
